Fix tetrahedron overlap checking algorithm
It's not enough to just check that the egdes of one tetrahedron don't intersect with the faces of the other; that check must be done in the other direction as well. So this is now done in two passes

Some of the optimizations done in the first pass don't apply to the second, so they are disabled during it. Later on, I may figure out some optimization(s) that can be done in the second pass.