[internal] Simplify the interface of TetrahedronOverlap
The change in 86b65fc4c9dcb69301bccd024f6129e583c7e7a8 was accompanied by an undocumented change in the TetrahedronOverlap interface: if its operator() returned true, setA() and setB() would both need to be called before calling operator() the next time. The program did do this, but it was a messy design.

It can now be called twice in a row in a conditiona when it returns "true", and return the same answer the second time, without setA() and setB() needing to be called first.
This change is just to illustrate that it could be done in this way, and will not be kept.