Cache branchless subpolytets that overlap
Only opposite ends of a candidate new polytet need to be actually tested for overlap now. The result is then cached, and subpolytets shorter than every candidate new polytet only need to be looked up in the overlap table.
This results in a huge speed boost, making the USE_GMP version almost as fast as the non-USE_GMP version.
First working implementation; a bit quick & dirty.