Minor speed boost by streamlining the detection of the 3-edge to 3-edge intersection case
The primary overlap-checking algorithm failed to detect an overlap if 3 edges of one tetrahedron intersected perfectly with 3 edges of another.
Instead of falling back on checking if the new vertex of tetrahedron A is inside tetrahedron B, count the number of times an edge of tetrahedron A intersects precisely with an edge of tetrahedron B, and this occurs twice for each face, treat that as an overlap.

This is an improvement upon aa0f4c6e6dcd558ca04e0064ed1007213baf449b.