[MULTITHREADING] Lock while reading the hash table
It's unclear whether this was actually responsible for any race conditions. Before this commit, some runs did have computation errors in term 17. After this commit, 5 runs experienced no errors, but then this commit was reverted as an experiment, and 7 further runs also experienced no errors.

In any case, the slowdown as a result of enabling this lock is only about 1.004×.